Move DotnetInfo interface and fix tests
